requirements.txt
numpy
epiweeks
torch
pandas
matplotlib
scipy
scikit-learn
PyYAML
pickle5
tqdm
statsmodels
dgl
hyperopt
properscoring
ipykernel